pyproject.toml
[project]
name = "sn-mcp"
version = "0.1.0"
description = "ServiceNow MCP utilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
sn-mcp = "server:main"
